Deep Convolutional Neural Networks/requirements.txt
Package                      Version
---------------------------- ------------
absl-py                      1.4.0
appdirs                      1.4.4
aspose-cells                 23.1.0
astunparse                   1.6.3
boto3                        1.26.76
botocore                     1.29.76
cachetools                   5.3.0
certifi                      2022.12.7
charset-normalizer           3.0.1
click                        8.1.3
colorama                     0.4.6
cssselect2                   0.7.0
cuda-python                  12.0.0
Cython                       0.29.33
deeplake                     3.2.9
dill                         0.3.6
entrypoints                  0.4
flatbuffers                  23.1.21
gast                         0.4.0
google-auth                  2.16.1
google-auth-oauthlib         0.4.6
google-pasta                 0.2.0
grpcio                       1.51.1
h5py                         3.8.0
hub                          3.0.1
humbug                       0.2.8
idna                         3.4
jmespath                     1.0.1
joblib                       1.2.0
JPype1                       1.4.1
keras                        2.11.0
libclang                     15.0.6.1
lxml                         4.9.2
Mako                         1.2.4
Markdown                     3.4.1
MarkupSafe                   2.1.2
multiprocess                 0.70.14
numcodecs                    0.11.0
numpy                        1.24.2
oauthlib                     3.2.2
opt-einsum                   3.3.0
packaging                    23.0
pandas                       1.5.3
pathos                       0.3.0
Pillow                       9.4.0
pip                          22.3.1
platformdirs                 3.0.0
pox                          0.3.2
ppft                         1.7.6.6
protobuf                     3.19.6
pyasn1                       0.4.8
pyasn1-modules               0.2.8
pycuda                       2022.2.2
PyJWT                        2.6.0
python-dateutil              2.8.2
pytools                      2022.1.14
pytz                         2022.7.1
reportlab                    3.6.12
requests                     2.28.2
requests-oauthlib            1.3.1
rsa                          4.9
s3transfer                   0.6.0
scikit-learn                 1.2.1
scipy                        1.10.1
setuptools                   65.5.0
six                          1.16.0
svg2rlg                      0.3
svglib                       1.5.1
tensorboard                  2.11.2
tensorboard-data-server      0.6.1
tensorboard-plugin-wit       1.8.1
tensorflow                   2.11.0
tensorflow-estimator         2.11.0
tensorflow-intel             2.11.0
tensorflow-io-gcs-filesystem 0.30.0
termcolor                    2.2.0
threadpoolctl                3.1.0
tinycss2                     1.2.1
torch                        1.13.1
torchaudio                   0.13.1+cu116
torchvision                  0.14.1+cu116
tqdm                         4.64.1
typing_extensions            4.5.0
urllib3                      1.26.14
webencodings                 0.5.1
Werkzeug                     2.2.3
wheel                        0.38.4
wrapt                        1.14.1
